+++ Sources/Superstring/AttributedStringBuildable.swift
@@ -24,25 +24,6 @@ public protocol AttributedStringBuildable {
     func applying(_ newAttributes: Attributes) -> Self
 }
 
-// MARK: - Ligature
-
-/// The value of this attribute is an integer.
-/// Ligatures cause specific character combinations to be rendered using a single custom glyph
-/// that corresponds to those characters.
-/// The default value for this attribute is 1. (Value 2 is unsupported on iOS).
-///
-/// - none: the value 0 indicates no ligatures.
-/// - default: this value indicates the use of the default ligatures.
-/// - all: this value indicates the use of all ligatures.
-public enum Ligature: Int {
-    case none = 0
-    case `default`   = 1
-
-    #if canImport(AppKit)
-    case all = 2 // Unsupported on iOS
-    #endif
-}
-
 extension AttributedStringBuildable {
 
     /************************ Attributes ************************/